+++ app/src/main/java/com/razdyakonov/movie_reviewer/MainActivity.kt
@@ -2,10 +2,19 @@ package com.razdyakonov.movie_reviewer
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.main_nav_host_fragment) as NavHostFragment
+        val navController = navHostFragment.navController
+        val bottomNavView = findViewById<BottomNavigationView>(R.id.bottom_nav_view)
+
+        bottomNavView?.setupWithNavController(navController)
     }
 }
