+++ app/src/main/java/com/razdyakonov/movie_reviewer/movie/favourites/FavouritesMoviesFragment.kt
@@ -1,0 +1,19 @@
+package com.razdyakonov.movie_reviewer.movie.favourites
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.razdyakonov.movie_reviewer.R
+
+class FavouritesMoviesFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_favourites_movies, container, false)
+    }
+}
