--- app/src/main/java/com/razdyakonov/movie_reviewer/movie/fragments/MovieViewModel.kt
@@ -1,7 +1,0 @@
-package com.razdyakonov.movie_reviewer.movie.fragments
-
-import androidx.lifecycle.ViewModel
-
-class MovieViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
