+++ app/src/main/java/com/razdyakonov/movie_reviewer/movie/searchmovies/MovieFragment.kt
@@ -1,0 +1,24 @@
+package com.razdyakonov.movie_reviewer.movie.searchmovies
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.razdyakonov.movie_reviewer.R
+
+class MovieFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.movie_fragment, container, false)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val viewModel = ViewModelProvider(this).get(MovieViewModel::class.java)
+    }
+}
