+++ app/src/main/java/com/razdyakonov/movie_reviewer/movie/searchmovies/MovieViewModel.kt
@@ -1,0 +1,7 @@
+package com.razdyakonov.movie_reviewer.movie.searchmovies
+
+import androidx.lifecycle.ViewModel
+
+class MovieViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
