+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/movieFragment"
+        android:icon="@drawable/ic_movie"
+        android:title="@string/bottom_nav_movies" />
+
+    <item
+        android:id="@+id/favouritesMoviesFragment"
+        android:icon="@drawable/ic_favourite"
+        android:title="@string/bottom_nav_movies" />
+
+</menu>
