+++ app/src/main/res/navigation/bottom_navigation_nav_graph.xml
@@ -7,7 +7,14 @@
 
     <fragment
         android:id="@+id/movieFragment"
-        android:name="com.razdyakonov.movie_reviewer.movie.fragments.MovieFragment"
+        android:name="com.razdyakonov.movie_reviewer.movie.searchmovies.MovieFragment"
         android:label="movie_fragment"
         tools:layout="@layout/movie_fragment" />
+
+    <fragment
+        android:id="@+id/favouritesMoviesFragment"
+        android:name="com.razdyakonov.movie_reviewer.movie.favourites.FavouritesMoviesFragment"
+        android:label="fragment_favourites_movies"
+        tools:layout="@layout/fragment_favourites_movies" />
+
 </navigation>
